split ops in ckernel_spfu into individual files
