+++ common/inc/sfpu/ckernel_sfpu_dropout.h
@@ -1,0 +1,94 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+template <bool APPROXIMATION_MODE, int ITERATIONS>
+inline void _calculate_dropout_(uint prob, uint scale)
+{
+    // SFPU microcode
+
+    FWLOG1("calculate_dropout() -- prob:%x", prob);
+    FWLOG1("calculate_dropout() -- scale:%x", scale);
+
+    vUInt rand = l_reg[LRegs::LReg3];
+    vUInt mask = reinterpret<vUInt>(vFloat(s2vFloat16b(0xa94b)));
+
+    #pragma GCC unroll 0
+    for (int d=0; d<4; d++) {
+        ////////////////////////
+        // Scale samples
+        ///////////////////////
+        dst_reg[0] = dst_reg[0] * s2vFloat16b(scale);
+
+        ////////////////////////
+        // Drop samples
+        ///////////////////////
+        vUInt tmp = rand >> 3;
+        v_if (tmp < vUInt(prob)) {
+            dst_reg[0] = vConst0;
+        }
+        v_endif;
+
+        ////////////////////////
+        // 16-bit PRNG update
+        ///////////////////////
+        tmp = rand << 1;
+
+        // Mask = 0x593CA -> 29e4d
+        // Mask = 0xd295 -> a94b
+        // PRNG SHL by one
+        v_if (tmp < 0) {
+            rand ^= mask;
+        }
+        v_endif;
+
+        dst_reg++;
+    }
+
+    l_reg[LRegs::LReg3] = rand;
+}
+
+inline void _init_dropout_seed_(uint16_t p2)
+{
+    FWLOG1("calculate_dropout() -- input seed:%x", p2);
+
+    uint32_t noc_id_reg = NOC_CMD_BUF_READ_REG(0, 0, NOC_NODE_ID);
+
+    uint16_t my_x = noc_id_reg & NOC_NODE_ID_MASK;
+    uint16_t my_y = (noc_id_reg >> NOC_ADDR_NODE_ID_BITS) & NOC_NODE_ID_MASK;
+
+    uint16_t per_tensix_input_seed = p2 ^ (my_x << my_y);
+
+    FWLOG1("calculate_dropout() -- calculated seed:%x", per_tensix_input_seed);
+
+    vInt result = l_reg[LRegs::LReg3];
+
+    vInt tmp = vConstTileId << 13;
+    vInt ptis = reinterpret<vInt>(vFloat(per_tensix_input_seed));
+    result = ~(tmp & ptis) & (tmp | ptis);
+
+    l_reg[LRegs::LReg3] = result;
+}
+
+inline void _init_dropout_(const uint seed)
+{
+    _init_dropout_seed_(seed);
+}
+
+} // namespace sfpu
+} // namespace ckernel
