+++ common/inc/sfpu/ckernel_sfpu_is_fp16_zero.h
@@ -1,0 +1,37 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+sfpi_inline vInt _sfpu_is_fp16_zero_(const vFloat& v, uint exponent_size_8)
+{
+    if (exponent_size_8) {
+        // fp16b
+        return v == 0.0F;
+    } else {
+        // fp16a
+        // if math data format is fp16, SFPU will convert 5 bit exp to 8 bit exp
+        // in grayskull, this unconditionally adds bias value to exp (even for zero)
+        vInt tmp = 0x3800; // loads {0, 8'd112, 10'b0}
+        tmp += reinterpret<vInt>(v);
+
+        return tmp == 0;
+    }
+}
+
+} // namespace sfpu
+} // namespace ckernel
