+++ common/inc/sfpu/ckernel_sfpu_power.h
@@ -1,0 +1,38 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ckernel_defs.h"
+#include "ckernel.h"
+#include "noc_nonblocking_api.h"
+
+#include "sfpi.h"
+
+using namespace sfpi;
+
+namespace ckernel
+{
+namespace sfpu
+{
+
+template <bool APPROXIMATION_MODE, int ITERATIONS>
+inline void _calculate_power_(uint exponent)
+{
+    for (int d = 0; d < ITERATIONS; d++)
+    {
+        vFloat in = dst_reg[0];
+        vFloat result = in * in;
+        for (uint i = 2; i < exponent; i++) {
+            result *= in;
+        }
+
+        dst_reg[0] = result;
+
+        dst_reg++;
+    }
+}
+
+} // namespace sfpu
+} // namespace ckernel
